Migrate to Vue SFC
