+++ package.json
@@ -44,6 +44,7 @@
     "tsc": "tsc -p . --noEmit",
     "test": "vitest run --project=unit --project=react --project=vue --project=solid",
     "lint": "eslint '{src,e2e}/**/*.{js,jsx,ts,tsx}'",
+    "check:vue": "vue-tsc --project ./tsconfig.vue.json",
     "check:svelte": "svelte-check --tsconfig ./tsconfig.svelte.json",
     "storybook": "storybook dev -p 6006",
     "storybook:rtl": "STORYBOOK_RTL=1 npm run storybook",
@@ -97,7 +98,6 @@
     "@types/react-virtualized": "^9.22.3",
     "@vitejs/plugin-react": "^5.0.4",
     "@vitejs/plugin-vue": "^6.0.1",
-    "@vitejs/plugin-vue-jsx": "^5.1.1",
     "@vitest/browser": "3.2.4",
     "@vitest/coverage-v8": "3.2.4",
     "babel-preset-solid": "^1.9.9",
@@ -134,13 +134,14 @@
     "typedoc-plugin-vue": "^1.5.0",
     "typescript": "^5.9.0",
     "typescript-eslint": "^8.43.0",
-    "unplugin-vue-jsx": "^0.8.0",
+    "unplugin-vue": "^7.0.2",
     "virtua": "^0.44.0",
     "vite": "^7.1.9",
     "vite-plugin-solid": "^2.11.9",
     "vitest": "^3.2.4",
-    "vue": "^3.4.34",
-    "vue-eslint-parser": "^10.2.0"
+    "vue": "^3.5.22",
+    "vue-eslint-parser": "^10.2.0",
+    "vue-tsc": "^3.1.1"
   },
   "peerDependencies": {
     "react": ">=16.14.0",
