+++ src/vue/index.ts
@@ -1,6 +1,9 @@
 /**
  * @module vue
  */
-export { VList } from "./VList.js";
-export { Virtualizer } from "./Virtualizer.js";
-export { WindowVirtualizer } from "./WindowVirtualizer.js";
+export { default as VList } from "./VList.vue";
+export type { VListHandle } from "./VList.vue";
+export { default as Virtualizer } from "./Virtualizer.vue";
+export type { VirtualizerHandle } from "./Virtualizer.vue";
+export { default as WindowVirtualizer } from "./WindowVirtualizer.vue";
+export type { WindowVirtualizerHandle } from "./WindowVirtualizer.vue";
